Let remove-from-head operation work
Check whether q is NULL or empty.
Check whether sp is non-NULL.If not,copy the removed string to sp,
with strncpy(), and set '\0' to the last byte of sp.
Use q_ele_ptr as temporary pointer to point the removed queue element.
Make the head of the queue point to the next of the removed queue
element.Finally, free the string and corresponding queue element,and
return true.