Let queue be able to be reversed
In q_reverse(),if q is not NULL and q is not empty,reverse is
available.At first, declare two pointers, ele_next and ele_now, to help
reverse the queue.The idae is that the function have a pointer go
through the old  queue list from the head to the tail.Before that
pointer goes to the next,the current pointed element is inserted in
queue from the head.ele_next is the moving pointer mentioned above,and
ele_now is used to point to the element to be inserted to the queue
from the head.Actually, ele_next begin going from the next of the
element pointed by the head of q.For the initial element to be the tail
of the reversed queue, NULL should be set to the next of the element.

In doing the above, the function need not allocate or free any list
element.